+++ src/Messenger/Transport/EmptyMessage.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AymDev\MessengerAzureBundle\Messenger\Transport;
+
+/**
+ * This message class is used when creating an envelope for decoding failures
+ * @internal
+ */
+final class EmptyMessage
+{
+}