+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2022-02-28
+### Added
+ - Converts the `MessageDecodingFailedException` to a new `SerializerDecodingException` containing an envelope with an empty message for logging purposes.
+
+### Fixed
+ - Deletes messages for consumers when the serializer throws a `MessageDecodingFailedException` to avoid retying them forever.
+
 ## [1.1.1] - 2022-02-15
 ### Fixed
  - **AzureBrokerPropertiesStamp** **DateTime** properties timezones are now set to the current default timezone.
@@ -21,7 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - **Symfony Messenger** transport for **Azure Service Bus** *queues* and *topics*
 
-[Unreleased]: https://github.com/AymDev/MessengerAzureBundle/compare/v1.1.1...HEAD
+[Unreleased]: https://github.com/AymDev/MessengerAzureBundle/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/AymDev/MessengerAzureBundle/releases/tag/v1.2.0
 [1.1.1]: https://github.com/AymDev/MessengerAzureBundle/releases/tag/v1.1.1
 [1.1.0]: https://github.com/AymDev/MessengerAzureBundle/releases/tag/v1.1.0
 [1.0.0]: https://github.com/AymDev/MessengerAzureBundle/releases/tag/v1.0.0